removed error

+++ neuron_network_lib.h
@@ -2,15 +2,15 @@
 
 typedef struct synapse 
 {
-  float *weight; 
-  float *value;
+  double *weight; 
+  double *value;
 } synapse;
 
 typedef struct output_neuron 
 {
   synapse **input;
   int input_count;
-  float activity;
+  double activity;
 } oneuron;
 
 typedef struct hidden_neuron 
@@ -19,17 +19,17 @@ typedef struct hidden_neuron
   synapse **output;
   int input_count;
   int output_count;
-  float activity;
+  double activity;
 } hneuron;
 
 typedef struct neural_network 
 {
   oneuron *oneurons; 
   hneuron **hneurons;
-  float *output_layer_weights;
-  float *output_layer_values; 
-  float **hidden_layer_weights; 
-  float **hidden_layer_values;
+  double *output_layer_weights;
+  double *output_layer_values; 
+  double **hidden_layer_weights; 
+  double **hidden_layer_values;
   unsigned int *hidden_layer_size;
   unsigned int hidden_layer_depth;
   unsigned int input_layer_size;
@@ -38,8 +38,8 @@ typedef struct neural_network
 
 neural_network *build_neural_network(unsigned int input_layer_size, 
   unsigned int *hidden_layer_size, unsigned int hidden_layer_depth, unsigned int output_layer_size);
-float *emulate(neural_network *nn, float *input);
-float adjust_weights(neural_network *nn, float *input, float *expected_output, float learning_rate);
-void teach(neural_network *nn, int number_of_samples, float *inputs, int number_of_inputs, float *expected_outputs, int number_of_outputs, float learning_rate, int passes);
-static inline float sigmoid(float x);
-static inline float random_weight(float min, float max);
+double *emulate(neural_network *nn, double *input);
+double adjust_weights(neural_network *nn, double *input, double *expected_output, double learning_rate);
+void teach(neural_network *nn, int number_of_samples, double *inputs, int number_of_inputs, double *expected_outputs, int number_of_outputs, double learning_rate, int passes);
+static inline double sigmoid(double x);
+static inline double random_weight(double min, double max);
